Compilation fix.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4633 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ monster.h
@@ -79,7 +79,7 @@ class Monster : public Creature
 		virtual bool isWalkable() const {return mType->isWalkable;}
 		virtual bool canSeeInvisibility() const {return Creature::isImmune(CONDITION_INVISIBLE);}
 		uint32_t getManaCost() const {return mType->manaCost;}
-		bool hasRecentBattle() const {return lastDamage && OTSYS_TIME() < (lastDamage + 30000);}
+		bool hasRecentBattle() const {return lastDamage && (uint64_t)OTSYS_TIME() < (lastDamage + 30000);}
 
 		void setSpawn(Spawn* _spawn) {spawn = _spawn;}
 		void setRaid(Raid* _raid) {raid = _raid;}
